Modified event redux and end now functionality
