+++ src/components/Calendar/Calendar.tsx
@@ -19,7 +19,6 @@ const Calendar = () => {
     displayEventTime={false}
     expandRows
   />
-
   );
 };
