+++ src/components/ViewMeeting/EndButton/EndButton.tsx
@@ -1,8 +1,10 @@
 import React, {useEffect} from "react";
 import Button from "@mui/material/Button";
 import {styled} from "@mui/material";
-import {deleteEvent} from "../../../redux/slices/eventSlice";
+import {deleteEvent} from "../../../store/Event/actionCreators";
 import {useDispatch} from "react-redux";
+import {FindUpcomingEvents} from "../../../utils/events.utils";
+import {useAppDispatch} from "../../../hooks/redux";
 
 const MeetingEndButton = styled(Button)({
   width: '100%',
@@ -17,15 +19,15 @@ const MeetingEndButton = styled(Button)({
   },
 })
 
-interface IsBusy {
-  isBusy: boolean
-}
+const EndButton = ({isBusy, upcomingEvent, getNextEventFunction}: any) => {
+  const dispatch = useAppDispatch();
 
-const EndButton = ({isBusy, upcomingEvent}: any) => {
-
-  const dispatch = useDispatch();
   const DeleteEvent = (id: number) => {
     dispatch(deleteEvent(id))
+      .unwrap()
+      .then(() => {
+        getNextEventFunction()
+      })
   }
   if (isBusy) {
     return (<MeetingEndButton onClick={() => DeleteEvent(upcomingEvent.elementId)}>End Now</MeetingEndButton>)
