--- src/interfaces/Events.ts
@@ -1,5 +1,0 @@
-import {IEvent} from "./Event";
-
-export interface IEvents {
-  events: Array<IEvent>
-}
