+++ src/interfaces/MeetingDetails.ts
@@ -1,8 +1,9 @@
-import {IEvent} from "./Event";
+import {IEvent} from "../models/Event";
 
 export interface MeetingDetails {
   isBusy: boolean,
-  upcomingEvent: IEvent,
+  upcomingEvent?: IEvent,
   seconds: number,
-  timeFunction(arg: boolean): void
+  timeFunction(arg: boolean): void,
+  getNextEventFunction?:() => void
 }
